Move some tests around and add some extra
